RAG/requirement.txt
transformers
torch
datasets
tqdm
rank_bm25
sentence-transformers
faiss-cpu
